refactor : ModalController 미사용하는 코드 제거
