+++ src/hooks/useModal/index.ts
@@ -1,3 +1,2 @@
-export { default as ModalController } from './ModalController';
 export { default as ModalProvider } from './ModalProvider';
 export { default as useModal } from './useModal';
